[CoreBundle][CustomerBundle] Move CustomerInitializer to CoreBundle, fix services.xml in UserBundle and CustomerBundle
