+++ src/Sylius/Bundle/CoreBundle/spec/Validator/Initializer/CustomerInitializerSpec.php
@@ -9,12 +9,12 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Sylius\Bundle\CustomerBundle\Validator\Initializer;
+namespace spec\Sylius\Bundle\CoreBundle\Validator\Initializer;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Sylius\Component\User\Canonicalizer\CanonicalizerInterface;
-use Sylius\Component\Customer\Model\CustomerInterface;
+use Sylius\Component\Core\Model\CustomerInterface;
 use Symfony\Component\Validator\ObjectInitializerInterface;
 
 /**
@@ -29,7 +29,7 @@ final class CustomerInitializerSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('Sylius\Bundle\CustomerBundle\Validator\Initializer\CustomerInitializer');
+        $this->shouldHaveType('Sylius\Bundle\CoreBundle\Validator\Initializer\CustomerInitializer');
     }
 
     function it_implements_symfony_validator_initializer_interface()
